restructured for setup
