+++ .gitignore
@@ -1,4 +1,23 @@
+# Distribution / packaging
+.Python
 *.pyc
+env/
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+*.egg-info/
+.installed.cfg
+*.egg
+
+# iac_athena
 *.idea/*
 .idea/
 *.log
@@ -6,3 +25,15 @@ lambda_package.zip
 .cache/
 __pycache__/
 .pytest_cache
+
+
+# Serverless directories
+.serverless
+
+htmlcov/
+.tox/
+.coverage
+cov_html/
+
+# ignore virtual environments used for testing
+venv/
