+++ boto_manager/__init__.py
@@ -1,0 +1,3 @@
+from .boto_client_manager import BotoClientManager
+
+__all__ = ['BotoClientManager', ]
