+++ boto_manager/tests/mock_data/sns.ListTopics_1.json
@@ -1,0 +1,25 @@
+{
+  "status_code": 200,
+  "data": {
+    "Topics": [
+      {
+        "TopicArn": "arn:aws:sns:us-west-2:000000000000:mock_test0"
+      },
+      {
+        "TopicArn": "arn:aws:sns:us-west-2:000000000000:mock_test1"
+      }
+    ],
+    "ResponseMetadata": {
+      "RetryAttempts": 0,
+      "HTTPStatusCode": 200,
+      "RequestId": "8d19e66b-d251-5590-bfee-b5dd3de9549a",
+      "HTTPHeaders": {
+        "x-amzn-requestid": "8d19e66b-d251-5590-bfee-b5dd3de9549a",
+        "vary": "Accept-Encoding",
+        "content-length": "10712",
+        "content-type": "text/xml",
+        "date": "Wed, 04 Jul 2018 15:53:27 GMT"
+      }
+    }
+  }
+}
