--- mock_data/sts.AssumeRole_1.json
@@ -1,35 +1,0 @@
-{
-    "status_code": 200, 
-    "data": {
-        "AssumedRoleUser": {
-            "AssumedRoleId": "000000000000000000000000000000:botocore-session-1530719606", 
-            "Arn": "arn:aws:sts::00000000000000000:assumed-role/iamr-sjd-role00000000000000000000/botocore-session-0000719606"
-        }, 
-        "Credentials": {
-            "SecretAccessKey": "00000000000000000000000000000000/", 
-            "SessionToken": "0000000000000000000000000000000/AeZz5ultK0s6y1SQlb9ZuKFDveClyvUqqIuEpsrlKj5m83Y7KPtiL6PNoIi/S2Qr92aECPD0S8eJCM1kdY+AbihHNPPBSc6sSvl3xqhULjdPjgcfVaMz31xnKBuMOqUMMyPF7WU/hwx43/oDzguGVGxWOOEgEV1QGfEl/SrS3CkZ/w+BzMHQgsCLrJqEIjo2CuLD8vxDnMvkGQFjMqtQpvWjl2b8Bv4SbPuwOUNLZnmqURU80xE6WhbvkQ5eefnjcGTWlEeg9rzSfeoEXu0f+bMdIzpCp64AivnTQwLMwj99QvTM/xujBTtm3fZOUNI18ExBrH9k1PAPLOVBrw8Hnyj32vPZBQ==", 
-            "Expiration": {
-                "hour": 16, 
-                "__class__": "datetime", 
-                "month": 7, 
-                "second": 27, 
-                "microsecond": 0, 
-                "year": 2018, 
-                "day": 3, 
-                "minute": 53
-            }, 
-            "AccessKeyId": "000000000000000000000000"
-        }, 
-        "ResponseMetadata": {
-            "RetryAttempts": 0, 
-            "HTTPStatusCode": 200, 
-            "RequestId": "642c1494-7fa2-11e8-8d04-d378ab94ee9b", 
-            "HTTPHeaders": {
-                "x-amzn-requestid": "642c1494-7fa2-11e8-8d04-d378ab94ee9b", 
-                "date": "Wed, 03 Jul 2018 15:53:27 GMT", 
-                "content-length": "1140", 
-                "content-type": "text/xml"
-            }
-        }
-    }
-}
