+++ requirements.txt
@@ -1,17 +1,4 @@
-atomicwrites==1.1.5
-attrs==18.1.0
-boto3==1.7.47
-botocore==1.10.47
-colorama==0.3.9
-docutils==0.14
-funcsigs==1.0.2
-futures==3.2.0
-jmespath==0.9.3
-more-itertools==4.2.0
-placebo==0.8.1
-pluggy==0.6.0
-py==1.5.4
-pytest==3.6.2
-python-dateutil==2.7.3
-s3transfer==0.1.13
-six==1.11.0
+pytest
+boto3
+placebo
+pytz
